Rework the information panel
Borrow the pqProxyInformationWidget from ParaView's Information Panel
to display information about the loaded volume. Hide some of the less
relevant info such as time steps, file size, number of points, number
of cells, and some info duplicate info such as volume bounds.

The main enhancement with this change is the display of the loaded
scalar arrays, their types, and their ranges in a table. Using
ParaView's UI infrastructure saves some rewriting of the same in
tomviz.