Add progress function to notify tomviz of updated data
This is rudimentary support for letting Python operators notify tomviz
that temporary data results are available.

Envisioned usage:

self.progress.data_updated()

On the C++ side, added to OperatorPython a member function
dataUpdated() that is invoked when the above Python function is
called. It tells the DataSource that the data has changed.

In addition, added a progress attribute for passing data from an
operator to the user interaction thread. Currently not working because
the pybind side is not happy with the VTK-Python-wrapped vtkImageData
it receives. We'll need to use an opaque pointer type to get around
this.